Update hello_world.ipynb
When you run this in some runtimes you get the following error:

WARNING: The TensorFlow contrib module will not be included in TensorFlow 2.0.
For more information, please see:
  * https://github.com/tensorflow/community/blob/master/rfcs/20180907-contrib-sunset.md
  * https://github.com/tensorflow/addons
If you depend on functionality not listed there, please file an issue.

Looking at the docs for TensorFlow it looks like the solution is to use the core impl of tf.enable_eager_execution instead of using the contrib lib. Looking at the TF core library it looks like this was moved to core ~1 year ago and so I would expect any runtimes supporting this to be capable of running that version of Tensor Flow.